+++ anybox/buildbot/openerp/build_utils/analyze_oerp_tests.py
@@ -9,16 +9,18 @@ import re
 FAILURE_REGEXPS = {
     'Failure in Python block': re.compile(r'WARNING:tests[.].*AssertionError'),
     'Errors during x/yml tests': re.compile(r'ERROR:tests[.]'),
-    'Errors or failures during unittest2 tests': re.compile(r'at least one error occurred in a test'),
+    'Errors or failures during unittest2 tests': re.compile(
+        r'at least one error occurred in a test'),
     'Errors loading addons': re.compile(r'ERROR.*openerp: Failed to load'),
     'Critical logs': re.compile(r'CRITICAL'),
     'Error init db': re.compile(r'Failed to initialize database'),
-    'Tests failed to excute': re.compile(r'openerp.modules.loading: Tests failed to execute'),
+    'Tests failed to excute': re.compile(
+        r'openerp.modules.loading: Tests failed to execute'),
 }
 
 test_log = open(sys.argv[1], 'r')
 
-failures = {} # label -> extracted line
+failures = {}  # label -> extracted line
 
 for line in test_log.readlines():
     for label, regexp in FAILURE_REGEXPS.items():
@@ -43,7 +45,3 @@ for label, failed_lines in failures.items():
 
 print "Total: %d failures " % total
 sys.exit(1)
-
-
-
-