+++ anybox/buildbot/openerp/build_utils/port_reserve.py
@@ -44,5 +44,3 @@ else:
 
 s.close()
 print(p)
-
-