+++ anybox/buildbot/openerp/buildouts.py
@@ -1,5 +1,6 @@
 from ConfigParser import ConfigParser
 
+
 def parse_manifest(filepath):
     """Return a ConfigParser for buildouts MANIFEST, ready for queries."""
     parser = ConfigParser()