+++ anybox/buildbot/openerp/tests/base.py
@@ -3,6 +3,7 @@ import unittest
 import shutil
 import tempfile
 
+
 class BaseTestCase(unittest.TestCase):
 
     def setUp(self):