+++ CHANGES.rst
@@ -1,6 +1,10 @@
 Changes
 ~~~~~~~
 
+0.9 (unreleased)
+----------------
+
+
 0.8
 ---
  - launchpad #1130838: build-only-if-requires buildslave option